Renamed WPEntity.java to EntityConfig.java
